+++ src/inmanta/server/services/compilerservice.py
@@ -139,11 +139,16 @@ class CompileRun(object):
         return ret
 
     async def get_branch(self) -> Optional[str]:
-        sub_process = await asyncio.create_subprocess_exec(
-            "git", "branch", stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd=self._project_dir
-        )
+        try:
+            sub_process = await asyncio.create_subprocess_exec(
+                "git", "branch", stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd=self._project_dir
+            )
 
-        out, err = await sub_process.communicate()
+            out, err = await sub_process.communicate()
+        finally:
+            if sub_process.returncode is None:
+                # The process is still running, kill it
+                sub_process.kill()
 
         if sub_process.returncode != 0:
             return None
@@ -163,7 +168,7 @@ class CompileRun(object):
                 env_all.update(env)
 
             sub_process = await asyncio.create_subprocess_exec(
-                cmd[0], *cmd[1:], stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd=cwd, env=env
+                cmd[0], *cmd[1:], stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd=cwd, env=env_all
             )
 
             returncode = await self.drain(sub_process)
@@ -174,6 +179,10 @@ class CompileRun(object):
         except Exception as e:
             await self._error("".join(traceback.format_exception(type(e), e, e.__traceback__)))
             return await self._end_stage(RETURNCODE_INTERNAL_ERROR)
+        finally:
+            if sub_process.returncode is None:
+                # The process is still running, kill it
+                sub_process.kill()
 
     async def run(self, force_update: Optional[bool] = False) -> Tuple[bool, Optional[model.CompileData]]:
         success = False
@@ -577,9 +586,6 @@ class CompilerService(ServerSlice):
     async def get_reports(
         self, env: data.Environment, start: Optional[str] = None, end: Optional[str] = None, limit: Optional[int] = None
     ) -> Apireturn:
-        argscount = len([x for x in [start, end, limit] if x is not None])
-        if argscount == 3:
-            return 500, {"message": "Limit, start and end can not be set together"}
         if env is None:
             return 404, {"message": "The given environment id does not exist!"}
 
@@ -594,9 +600,20 @@ class CompilerService(ServerSlice):
             start_time = dateutil.parser.parse(start)
         if end is not None:
             end_time = dateutil.parser.parse(end)
-        models = await data.Compile.get_reports(env.id, limit, start_time, end_time)
 
-        return 200, {"reports": models}
+        models = await data.Compile.get_list_paged(
+            page_by_column="started",
+            order_by_column="started",
+            order="DESC",
+            limit=limit,
+            start=start_time,
+            end=end_time,
+            no_obj=False,
+            connection=None,
+            environment=env.id,
+        )
+
+        return 200, {"reports": [m.to_dict() for m in models]}
 
     @protocol.handle(methods.get_report, compile_id="id")
     async def get_report(self, compile_id: uuid.UUID) -> Apireturn: