+++ CHANGELOG.md
@@ -5,11 +5,21 @@
 - Added more specific location information for attributes (#2481)
 - Added plugin call anchors to support ctrl-clicking a plugin call (#1954)
 - Added rpdb signal handler (#2170)
+- Added pagination support on api calls for agent and agentproc (#2500)
+- Added support to build RPMs for a python version different from Python3.6 (#1857)
+- Added support for assigning `null` to relations with lower arity 0 (#2459)
 - Added documention on the core dashboard (#dashboard-63)
 
 ## Bug fixes
 - Fix broken links in the documentation (#2495)
 - Ensure all running compilations are stopped when the server is stopped (#2508)
+- Cleanup old entries in the agentprocess and agentinstance database tables (#2499)
+- Ensure the compiler service takes into account the environment variables set on the system (#2413)
+- Fix `--server_address` option on `inmanta export` (#2514)
+- Handle failure in an event handler consistently for local and non-local agents (#2509)
+- Fix for cross agent dependencies responding to unavailable resources (#2501)
+- Handle JSON serialization errors in handler log messages (#1875)
+- Export command should raise exception on failure (#2487)
 
 ## Upgrade notes
 - Ensure the database is backed up before executing an upgrade.